+++ lib/src/format/birthday.dart
@@ -1,5 +1,5 @@
+import '../constant/date.dart';
 import 'package:intl/intl.dart';
-import 'package:hive_dart/src/constant/date.dart';
 
 /// 把时间戳格式化为 10.01 格式
 String formatBirthday(int value) {