+++ lib/src/format/month.dart
@@ -1,5 +1,5 @@
 import 'package:intl/intl.dart';
-import 'package:hive_dart/src/constant/month.dart';
+import '../constant/month.dart';
 
 /// 把时间戳格式化为 2020-10 格式
 String formatMonth(int value) {