+++ lib/src/format/year.dart
@@ -1,5 +1,5 @@
 import 'package:intl/intl.dart';
-import 'package:hive_dart/src/constant/year.dart';
+import '../constant/year.dart';
 
 /// 把时间戳格式化为 2020 格式
 String formatYear(int timestamp) {