+++ lib/src/util/time.dart
@@ -1,6 +1,5 @@
-import 'package:hive_dart/hive_dart.dart';
 import 'package:intl/intl.dart';
-
+import '../constant/millisecond.dart';
 
 /// 时间戳转成时间对象
 DateTime? parseTime(String str, String format) {