+++ CHANGELOG.md
@@ -95,9 +95,9 @@
 - 新增 randomIntegerByLength, randomIntegerByRange, randomStringByLength, randomStringByCurrentTime
 - 新增 normalizeUrl, toProtocolRelativeUrl, encodeUriComponent, decodeUriComponent
 - 新增 darkenColor, lightenColor, hasSpecialCharacters, removeSpecialCharacters
-- 新增 parseInteger、parseNumber、applyRate
+- 新增 parseInteger, parseNumber, applyRate
 
 ## 0.2.3
 
-- 新增 isCorporateAccountNumber, isMobile, hasDecimal, parseTime
-- 删除 isInteger, stringToTime, formatProvince, formatCity, formatDistrict, formatHourMinutes
+- 新增 isCorporateAccountNumber, isMobile, hasDecimal, parseTime, hasSpecialCharacter, removeSpecialCharacter, applyRateFloor, applyRateCeil, applyRateRound, toHttpProtocolUrl, toRelativeProtocolUrl
+- 删除 isInteger, stringToTime, formatProvince, formatCity, formatDistrict, formatHourMinutes, hasSpecialCharacters, removeSpecialCharacters, applyRate, normalizeUrl, toProtocolRelativeUrl